pyproject.toml
[project]
name = "scrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "apache-airflow[celery]>=3.1.7",
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pandas>=3.0.1",
    "pyspark>=4.1.1",
    "python-dotenv>=1.2.1",
]
